+++ components/front_matter/src/page.rs
@@ -33,6 +33,7 @@ pub struct PageFrontMatter {
     /// Integer to use to order content. Highest is at the bottom, lowest first
     pub weight: Option<usize>,
     /// All aliases for that page. Gutenberg will create HTML templates that will
+    /// redirect to this
     #[serde(skip_serializing)]
     pub aliases: Option<Vec<String>>,
     /// Specify a template different from `page.html` to use for that page